django-task-mgmt/requirements.txt
Django==1.6.2
django-tastypie==0.10.0
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.3
static==1.0.2
wsgiref==0.1.2
